Backend/villagr/requirements.txt
Django
django-phone-field
djangorestframework
drf-extensions
markdown